Add files via upload
GUI is complete, and all but 1 minor issue I'm having, the code is pretty much complete.  The issue is that either child window can be called without restriction.  I would like to be able to disable the buttons while a child window is open and then enable them when the window closes, but I am having issues with passing the methods between classes.  fix will be coming.